requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==4.0.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
bcrypt==4.2.0
beautifulsoup4==4.12.2
bidict==0.23.1
bleach==6.1.0
blinker==1.7.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1677499911949/work
branca==0.7.2
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1693583518054/work
cachetools==5.3.2
certifi @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_3bweqb8byn/croot/certifi_1690232244779/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1671179893800/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1688813409104/work
click==8.1.7
cloudpickle==3.0.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm==0.2.0
conda==23.3.1
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1680508672016/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1691048088238/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1691009212940/work
contourpy==1.2.0
cryptography @ file:///Users/runner/miniforge3/conda-bld/cryptography-split_1691444278572/work
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.0.4
efficientnet @ file:///Users/nisha/Desktop/spring2024/ResponsibleDataScience/final_project/input/efficientnet/efficientnet-1.1.0
exceptiongroup==1.1.3
executing==2.0.1
faicons==0.2.2
fairlearn==0.10.0
fastapi==0.110.0
fastjsonschema==2.19.0
filelock==3.14.0
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==23.5.26
fonttools==4.47.2
fqdn==1.5.1
fsspec==2024.3.1
gast==0.5.4
GeoAlchemy2==0.15.2
geographiclib==2.0
geopy==2.4.1
google-api-core==2.18.0
google-auth==2.29.0
google-auth-oauthlib==1.1.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
graphviz==0.20.1
grpcio==1.59.3
gunicorn==23.0.0
h11==0.14.0
h5py @ file:///Users/ktietz/ci_310/h5py_1644262287129/work
htmltools==0.5.1
huggingface-hub==0.23.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
image==1.5.33
imageio==2.34.1
imbalanced-learn==0.12.0
ipykernel==6.26.0
ipyleaflet==0.19.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonify==0.5
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1632759296524/work
jsonpointer==2.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-leaflet==0.19.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
keras==2.15.0
Keras-Applications @ file:///Users/nisha/Desktop/spring2024/ResponsibleDataScience/final_project/input/kerasapplications/keras-team-keras-applications-3b180cb
kiwisolver==1.4.5
labelImg==1.8.6
lazy_loader==0.4
libclang==16.0.6
libmambapy @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1680791351028/work/libmambapy
linkify-it-py==2.0.3
llvmlite==0.42.0
lxml==5.2.1
mamba @ file:///Users/runner/miniforge3/conda-bld/mamba-split_1680791351028/work/mamba
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.2.0
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.3
nltk==3.8.1
node==1.2.1
notebook==7.0.6
notebook_shim==0.2.3
numba==0.59.1
numpy==1.26.4
oauthlib==3.2.2
odict==1.9.0
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
overrides==7.4.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas==2.2.0
pandocfilters==1.5.0
parso==0.8.3
patchify==0.2.3
patsy==0.5.6
pexpect==4.8.0
pillow==10.2.0
platformdirs==4.0.0
plotly==5.22.0
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1693086607691/work
plumber==1.7
prometheus-client==0.18.0
prompt-toolkit==3.0.36
proto-plus==1.23.0
protobuf==4.23.4
psutil==5.9.6
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==15.0.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1666836623787/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic==2.6.3
pydantic_core==2.16.3
pydicom==2.4.4
pygame==2.5.2
Pygments==2.16.1
PyMySQL==1.1.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1685514481738/work
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.14
PyQt5-sip==12.13.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
pyspark==3.5.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-json-logger==2.0.7
python-multipart==0.0.9
python-socketio==5.11.2
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
questionary==2.0.1
referencing==0.31.0
regex==2023.10.3
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ridgeplot==0.1.25
rpds-py==0.13.0
rsa==4.9
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1686993977949/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1670412881259/work
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.12.2
Send2Trash==1.8.2
shap==0.45.1
shiny==0.9.0
shinyswatch==0.6.0
shinywidgets==0.3.2
simple-websocket==1.0.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
slicer==0.0.8
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.36.3
statsmodels==0.14.2
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.3.0
terminado==0.18.0
threadpoolctl==3.2.0
tifffile==2024.5.3
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
torch==2.3.0
torchvision==0.18.0
tornado==6.3.3
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1691671248568/work
traitlets==5.13.0
traittypes==0.2.1
transformers==4.40.1
typeguard==2.13.3
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.4
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1689789803562/work
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.1.2
xyzservices==2024.4.0
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==6.3
zope.lifecycleevent==5.0
zope.proxy==5.2
zstandard @ file:///Users/runner/miniforge3/conda-bld/zstandard_1667296287896/work
